move image widget into new file
